+++ src/plugins/vue-i18n.js
@@ -1,4 +1,4 @@
-// we import the external package
+import Vue from 'vue';
 import VueI18n from 'vue-i18n';
 
 // let's say we have a file in /src/i18n containing the language pack
@@ -7,28 +7,36 @@ import messages from 'src/i18n';
 import { Cookies } from 'quasar';
 import { WEB_CONSTANTS } from 'shared/Constants';
 
-export default ({ app, Vue }) => {
-  // we tell Vue to use our Vue package:
-  Vue.use(VueI18n);
+let klabI18N = null;
 
-  // Set i18n instance on app;
-  // We inject it into root component by doing so;
-  // new Vue({..., i18n: ... }).$mount(...)
-  let cookieLocale = 'en';
-  if (Cookies.has(WEB_CONSTANTS.COOKIE_LANG)) {
-    cookieLocale = Cookies.get(WEB_CONSTANTS.COOKIE_LANG);
-    console.debug(`Locale setted from cookie to ${cookieLocale}`);
-  } else {
-    Cookies.set(WEB_CONSTANTS.COOKIE_LANG, cookieLocale, {
-      expires: 30,
-      path: '/',
+export function getI18N(reload = false) {
+  if (reload || klabI18N === null) {
+    Vue.use(VueI18n);
+
+    // Set i18n instance on app;
+    // We inject it into root component by doing so;
+    // new Vue({..., i18n: ... }).$mount(...)
+    let cookieLocale = 'en';
+    if (Cookies.has(WEB_CONSTANTS.COOKIE_LANG)) {
+      cookieLocale = Cookies.get(WEB_CONSTANTS.COOKIE_LANG);
+      console.debug(`Locale setted from cookie to ${cookieLocale}`);
+    } else {
+      Cookies.set(WEB_CONSTANTS.COOKIE_LANG, cookieLocale, {
+        expires: 30,
+        path: '/',
+      });
+      console.debug(`Lang cookie setted to ${cookieLocale}`);
+    }
+    klabI18N = new VueI18n({
+      locale: cookieLocale,
+      fallbackLocale: 'en',
+      messages,
     });
-    console.debug(`Lang cookie setted to ${cookieLocale}`);
   }
+  return klabI18N;
+}
 
-  app.i18n = new VueI18n({
-    locale: cookieLocale,
-    fallbackLocale: 'en',
-    messages,
-  });
+export default ({ app }) => {
+  // we tell Vue to use our Vue package:
+  app.i18n = getI18N();
 };