+++ src/shared/Constants.js
@@ -1,6 +1,8 @@
 /**
  * Constants of application
  */
+import { getI18N } from 'plugins/vue-i18n';
+
 export const CONSTANTS = {
   /**
    * Used to centralize the default empty map selection
@@ -417,7 +419,7 @@ export const WEB_CONSTANTS = {
   COOKIE_MAPDEFAULT: 'klab_mapdefault',
   COOKIE_SAVELOCATION: 'klab_saveLocation',
   COOKIE_HELP_ON_START: 'klab_helponstart',
-  COOKIE_DOCKED_STATUS: 'klab_dockedstatus', // will be true | false | not setted
+  COOKIE_DOCKED_STATUS: 'klab_dockedstatus', // will be true | false | not set
 };
 
 export const COLORS = {
@@ -714,6 +716,30 @@ export const HELP_CONSTANTS = {
   DEFAULT_HELP_BASE_URL: 'http://www.integratedmodelling.org/statics/help',
 };
 
+export const OBSERVATION_CONTEXT_EMPTY_ITEM = {
+  actionLabel: null,
+  actionId: null,
+  downloadUrl: null,
+  downloadFileExtension: null,
+  enabled: false,
+  separator: false,
+  submenu: [],
+};
+
+export const OBSERVATION_CONTEXT_ITEMS = {
+  SEPARATOR_ITEM: {
+    ...OBSERVATION_CONTEXT_EMPTY_ITEM,
+    enabled: true,
+    separator: true,
+  },
+  RECONTEXTUALIZATION_ITEM: {
+    ...OBSERVATION_CONTEXT_EMPTY_ITEM,
+    actionId: 'Recontextualization',
+    actionLabel: getI18N().tc('label.recontextualization'),
+    enabled: true,
+  },
+};
+
 export default {
   CONSTANTS,
   LEFTMENU_CONSTANTS,
@@ -741,4 +767,5 @@ export default {
   MODIFICATIONS_TYPE,
   SETTING_NAMES,
   HELP_CONSTANTS,
+  OBSERVATION_CONTEXT_ITEMS,
 };