+++ src/store/view/state.js
@@ -146,8 +146,11 @@ export default {
   treeSelected: null,
   treeTicked: [],
   treeExpanded: [],
+  /**
+   * The raster layer on the top
+   * Vector layer are not used
+  */
   topLayer: null,
-  loadingLayers: [],
   /**
    * true if we are editing the scale. Need for no intercept events
    */